Move the todo reducer to a custom reducer file
